Refactor: move Contribution.kt to data.models
