+++ app/src/main/java/fr/free/nrw/commons/contributions/WikipediaInstructionsDialogFragment.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.view.WindowManager
 import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
+import fr.free.nrw.commons.data.models.Contribution
 import fr.free.nrw.commons.databinding.DialogAddToWikipediaInstructionsBinding
 
 /**
